refactor(codec): move JB2/IW44 encoders into their crates and share spec data (#344)
Both bilevel (JB2) and wavelet (IW44) encoders lived in the monolith as
`src/jb2_encode.rs` / `src/iw44_encode.rs`, duplicating spec tables and
context state that the decoder already owns in the codec crates. Follow
the `djvu-bzz` model: gate a `pub mod encode` behind `std` in each crate
and move the encoder next to its decoder, with one source of truth for
the shared spec data.

JB2 (djvu-jb2):
* Add `#[cfg(feature = "std")] pub mod encode;`; `git mv` the encoder in.
* Promote the shared NumContext / state-flag / record-type definitions to
  `pub(crate)`; delete the encoder's private copies.
* Gate the experiment-only `CrossSizeRec6Probe` type and the
  `Jb2EncodeOptions::cross_size_rec6_probe` field behind a new
  `experimental` feature so they leave the published interface (#322/#301
  keep their entry points; only the *types* move off the public surface).

IW44 (djvu-iw44):
* Add `#[cfg(feature = "std")] pub mod encode;`; `git mv` the encoder in.
* Promote the band/quant init tables, coefficient state flags, and the
  zigzag bit-interleave const fns to `pub(crate)` (one spec shared by the
  decoder and the std-only encoder); delete the duplicated copies. The
  encoder keeps its forward `ZIGZAG_ROW`/`COL` tables (encoder-specific).
* Move the reconstruction-loss diagnostics (originally #320) into the
  crate's `#[cfg(test)]` module, re-based on a deterministic synthetic 1/f
  octave-sinusoid image instead of a decoded corpus page (the corpus path
  needed the monolith's `DjVuDocument` parser, absent from the codec
  crate). Adds zero public/experimental API surface.

Monolith:
* `src/jb2_encode.rs` / `src/iw44_encode.rs` become thin re-export shims
  (`pub use djvu_*::encode::*;`). JB2's shim keeps the DJVM multi-page
  bundling helpers and the #322 corpus measurement driver, which compose
  the container layer and so cannot live in the codec crate.
* Wire a top-level `experimental` feature → `djvu-jb2/experimental`, and
  gate the `encode_quality_djbz` example on it.
* The `encode_quality_jb2` example now spreads `..Default::default()`
  under an `#[allow(clippy::needless_update)]`: needless without
  `experimental` (one field), required with it (two).

No_std decode builds for both crates still pass; encode/decode round-trip
and band-residual property tests are unchanged in behavior.
